Translation in French
Hello,

I translated the config.json in english to french. I performed it using Chummer5a, Shadowrun 5e PDF in english and in french. It might not be perfect but i will upgrade it when I'll play my new Shadowrun campaign using FoundryVTT.
I'll upload a new request in the system.json. 
Also, do you need me to pull a request to add the config.json in the folder SR5-foundryVTT/locale? I don't know the utility of this folder.

Quentin