Treat condition monitors like health bars
The token resource bars are designed to function like health bars, with
a full status being good and an empty status being bad. They are
color-coded accordingly with, empty being red and full being green on
the primary resource bar.
The Shadowrun condition monitor works oppositely, however, with full
being bad and empty being full.

This commit flips the value of the token's copy of the condition
monitor resources just before they are drawn so they draw correctly
when drawn as health bars.