[CODE] Implemented FoundryVTT v11 nedb devops mechanism (#1204)
Instead of including the previous LevelDB .db-Files, we commit source document json files directly to git and build a new compendium on the fly for release and dev builds.

For this we include the dnd5e mechanism of wrapping foundryvtt-cli methods through utils/packs.mjs and a new npm run build:db script. See readme-dev.md and utils/packs.mjs for more details.