Additional Changes
I have implemented the following changes after consultation with the Discord:

Yellow drop shadow to more clearly delineate item and actors sheets.

removed the heading from actor and item sheets and minimised the icons to be less invasive but still functional.

The Action tab on items has seen some changes, I have reduced the size of the headings and made them yellow. They were rather large before and this keeps the visible as headings but less visually obtrusive.

I have altered the journal title area. This is now flush to the edges.

These changes are all easily reversable if

a) the discord channel don't like after seeng it in action or
b) if not favourably recieved by players

once this is all committed I will move on to light and dark modes